Feature: Create new plant

+++ src/main/java/daos/GenericDaoInterface.java
@@ -7,7 +7,7 @@ import java.util.List;
 public interface GenericDaoInterface<T> {
     List<T> executeQuery(String sql, RowMapperInterface<T> rowMapper, Object... parameters);
 
-    Long executeUpdate(String sql, Object... parameters);
+    Integer executeUpdate(String sql, Object... parameters);
 
     int count(String sql, Object... parameters);
 }
